17-04-22

+++ src/main/java/com/example/ordersdelivery/dto/RouteDetailsDTO.java
@@ -11,5 +11,5 @@ public interface RouteDetailsDTO {
     int getQty();
     Long getTransportId();
     String getTransportType();
-    Long getTransportVolume();
+    double getTransportVolume();
 }
